Remove lantern_ prefix from cli modules
