+++ lantern_cli/src/daemon/migrations.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
-use crate::lantern_logger::Logger;
-use crate::lantern_utils::get_full_table_name;
+use crate::logger::Logger;
+use crate::utils::get_full_table_name;
 use tokio_postgres::{Client, NoTls};
 
 use super::cli;
@@ -14,8 +14,8 @@ pub async fn drop_old_triggers_and_functions(
     client
         .batch_execute(&format!(
             "
-       DROP FUNCTION IF EXISTS notify_insert_lantern_daemon CASCADE;
-       DROP FUNCTION IF EXISTS notify_update_lantern_daemon CASCADE;
+       DROP FUNCTION IF EXISTS notify_insert_daemon CASCADE;
+       DROP FUNCTION IF EXISTS notify_update_daemon CASCADE;
        DROP TRIGGER IF EXISTS trigger_lantern_jobs_insert ON {emb_jobs_full_table_name} CASCADE;
        DROP TRIGGER IF EXISTS trigger_lantern_jobs_update ON {emb_jobs_full_table_name} CASCADE;
     ",
