+++ lantern_cli/src/embeddings/cli.rs
