+++ lantern_cli/src/embeddings/core/cohere_runtime.rs
