+++ lantern_cli/src/embeddings/core/mod.rs
