+++ lantern_cli/src/embeddings/core/openai_runtime.rs
