+++ lantern_cli/src/embeddings/core/utils.rs
