+++ lantern_cli/src/external_index/cli.rs
