+++ lantern_cli/src/external_index/utils.rs
