+++ lantern_cli/src/lib.rs
@@ -1,10 +1,10 @@
-pub mod lantern_daemon;
-pub mod lantern_embeddings;
-pub mod lantern_external_index;
-pub mod lantern_index_autotune;
-pub mod lantern_logger;
-pub mod lantern_pq;
-pub mod lantern_utils;
+pub mod daemon;
+pub mod embeddings;
+pub mod external_index;
+pub mod index_autotune;
+pub mod logger;
+pub mod pq;
+pub mod utils;
 pub mod types;
 
 #[macro_use]
