+++ lantern_cli/src/pq/gcp_batch.rs
@@ -2,8 +2,8 @@ use super::cli::PQArgs;
 use super::setup::{make_codebook_logged_and_readonly, setup_tables, setup_triggers};
 use super::{set_and_report_progress, AnyhowVoidResult, ProgressCbFn};
 use isahc::{prelude::*, HttpClient, Request};
-use crate::lantern_logger::Logger;
-use crate::lantern_utils::quote_ident;
+use crate::logger::Logger;
+use crate::utils::quote_ident;
 use postgres::{Client, NoTls};
 use serde::Deserialize;
 use serde_json::{self, json, Value};
